requirements.txt
# Fundamental
numpy==2.1.2
jax==0.6.2
jaxlib==0.6.2
optax==0.2.6
flax==0.10.7
dm-haiku==0.0.15
numpyro==0.19.0
numba==0.62.1
scipy==1.15.3

# Logging / visualization
tensorboardX==2.6.4
tensorboard==2.20.0
tbparse==0.0.9
matplotlib==3.10.7
seaborn==0.13.2
pandas==2.3.3
wandb==0.22.3

# Used by environment
gymnasium[mujoco]==1.0.0
brax==0.14.0
dill==0.4.0
mujoco==3.4.0

# Utility
tqdm==4.67.1
setproctitle==1.3.7
pyyaml==6.0.3
